Organize top nav files
