+++ src/serverfns.rs
@@ -2,6 +2,6 @@ pub mod auth;
 pub mod comments;
 pub mod communities;
 pub mod get_site;
-pub mod get_theme;
 pub mod posts;
+pub mod theme;
 pub mod users;
